fix: improve CBIR_Texture and CBIR Colour Runtime. feat: add json files and dataset

+++ src/function/ImageToMatrix.py
@@ -1,9 +1,10 @@
 import cv2
 import numpy as np
+import cosine_similiarity as cs
 
 def ImageToMatrix(path):
     img_bgr = cv2.imread(path)
-    img_rgb = cv2.cvtColor(img_bgr, cv2.COLOR_BGR2RGB).astype('float64')
+    img_rgb = cv2.cvtColor(img_bgr, cv2.COLOR_BGR2RGB).astype('float32')
     return img_rgb
 
 def RGBtoHSV(rgb_matrix):
@@ -17,7 +18,7 @@ def RGBtoHSV(rgb_matrix):
     h = np.where(Delta == 0, 0,
                 np.where(Cmax == r, 60 * ((g - b) / Delta % 6), #Kasus Cmax = r
                 np.where(Cmax == g, 60 * ((b - r) / Delta + 2), #Kasus Cmax = g
-                60 * ((r - g) / Delta + 4)))) #Kasus Cmax = b
+                60 * ((r - g) / Delta + 4))))                   #Kasus Cmax = b
 
     s = np.where(Cmax == 0, 0, Delta / Cmax)
     v = Cmax
